% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Utils.R
\name{COUNT2TPM}
\alias{COUNT2TPM}
\title{Title Transform a count table to TPM matrix}
\usage{
COUNT2TPM(COUNT)
}
\arguments{
\item{COUNT}{a dataframe with each row represents a gene and each column represents a cell}
}
\value{
TPM Dataframe
}
\description{
Title Transform a count table to TPM matrix
}
\examples{
tpm <- COUNT2TPM(count)
}
