% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Clean.R
\name{DeleteColsByNames}
\alias{DeleteColsByNames}
\title{Title Delete rows by names of a dataframe}
\usage{
DeleteColsByNames(DF, col.names.remove)
}
\arguments{
\item{DF}{dataframe}

\item{col.names.remove}{names of col to be removed}
}
\value{
dataframe
}
\description{
Title Delete rows by names of a dataframe
}
\examples{
clean_cexpr <- DeleteColsByNames(expr,c("Lars2","Gm10800"))
}
