% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Clean.R
\name{DeleteRowsByNames}
\alias{DeleteRowsByNames}
\title{Title Delete rows by names of a dataframe}
\usage{
DeleteRowsByNames(DF, row.names.remove)
}
\arguments{
\item{DF}{dataframe}

\item{row.names.remove}{names of rows to be removed}
}
\value{
dataframe
}
\description{
Title Delete rows by names of a dataframe
}
\examples{
clean_cexpr <- DeleteRowsByNames(expr,c("Lars2","Gm10800"))
}
