% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Utils.R
\name{Desq2}
\alias{Desq2}
\title{Title}
\usage{
Desq2(count1, count2, group_pattern, thred = 20)
}
\arguments{
\item{count1}{Dataframe. Count table 1, with each row is a gene and each col is a cell.}

\item{count2}{Dataframe. Count table 2, with each row is a gene and each col is a cell.}

\item{group_pattern}{list. Group label usually a shared substr of colnames(count), eg c("RAW24","RAW48")}

\item{thred}{int(default 20)}
}
\value{
dataframe with DE genes (padj, FoldChange)
}
\description{
Title
}
\examples{
de24_48 <- Desq2(RAW24h_count,RAW48h_count,c("RAW24","RAW48"),thred = 20)
head(rownames(de24_48),n=30)
}
