% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Plot.R
\name{Embedding}
\alias{Embedding}
\title{Title Dimension reduction plot (2D) with tsne/umap}
\usage{
Embedding(exprmatx, method = "umap", rdm = 123, n_neighours = 20,
  min_dist = 0.5, perplex = 25, iter = 8000)
}
\arguments{
\item{exprmatx}{Dataframe dataframe with each row represents a gene and each column represents a cell}

\item{method}{str (default tsne). "umap" / "tsne"}

\item{rdm}{int. (default 123) random seed number for umap (https://umap-learn.readthedocs.io/en/latest/parameters.html).}

\item{n_neighours}{int. (default=20) param for umap. Higher, more global structure .}

\item{min_dist}{float. (default=0.1) between 0-1, Higher, more global structure.}

\item{perplex}{int. default = 25 tsne para.}

\item{iter}{int. default = 8000 tsne para.}
}
\value{
a ggplot object
}
\description{
Title Dimension reduction plot (2D) with tsne/umap
}
\examples{
p <- Embedding(expr,"umap")
}
