% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Clean.R
\name{FilterCellBySumCut}
\alias{FilterCellBySumCut}
\title{Title Quality Control on cells by summary matrix}
\usage{
FilterCellBySumCut(summary_, uniqmp = u, prgene = p, mt = m,
  ercc = e, rRNA = r)
}
\arguments{
\item{summary_}{Str, path to summary file (with header and rownames)}

\item{uniqmp}{list of length two. u[1] is the lower bound, and u[2] is the upper bound.}

\item{prgene}{list of length two.}

\item{mt}{list of length two.}

\item{ercc}{list of length two.}

\item{rRNA}{list of length two.}
}
\value{
list, list of filtered cell names.
}
\description{
Title Quality Control on cells by summary matrix
}
\examples{
filtered_cells <- FilterCellBySumCut(sum,uniqmp=c(60,100),prgene=c(4000,12000),mt=c(0,10),ercc=c(0,15,rRNA=c(0,20)))
}
