% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Clean.R
\name{FilterGenesExpr}
\alias{FilterGenesExpr}
\title{Title Filter out genes with expression level below threshold (thre)
if the max expression level of gene is below thre, then drop the gene.}
\usage{
FilterGenesExpr(DF, thre = 3)
}
\arguments{
\item{DF}{dataframe with each row represents a gene and each column represents a cell}

\item{thre}{expression levele threshold}
}
\value{
A dataframe
}
\description{
Title Filter out genes with expression level below threshold (thre)
if the max expression level of gene is below thre, then drop the gene.
}
\examples{
high_gene_expr_mtx <- FilterGenesExpr(matx,thre=5)
}
