% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Clean.R
\name{FilterGenesMinCells}
\alias{FilterGenesMinCells}
\title{Title Delete genes expressed lower than min_cell.
if min_cell=0, then it will delete genes with zero expression level across all cells.}
\usage{
FilterGenesMinCells(count_table, min_cell = 3)
}
\arguments{
\item{count_table}{A dataframe with each row represents a gene and each column represents a cell}

\item{min_cell}{int (default 3), minimum number of cells that a gene manifests.}
}
\value{
Filtered single cell expressed profile.
}
\description{
Title Delete genes expressed lower than min_cell.
if min_cell=0, then it will delete genes with zero expression level across all cells.
}
