% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Plot.R
\name{Plot_Embed_Category}
\alias{Plot_Embed_Category}
\title{Title Plot 2D embedding with factor group as color scale}
\usage{
Plot_Embed_Category(exprmatx, embd, group = "red", title = "")
}
\arguments{
\item{exprmatx}{Dataframe dataframe with each row represents a gene and each column represents a cell}

\item{embd}{datafram. By calling Embedding(exprmatx).}

\item{group}{list. (default="red") corresponding to colnames of exprmatx}

\item{title}{str. Fig title.}
}
\value{
a ggplot object
}
\description{
Title Plot 2D embedding with factor group as color scale
}
\examples{
sti_vs_ctrl <- Plot_Embed_Category(exprmatx,embd,roup=c(rep("sti",3),rep("ctrl",5)))
}
