% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Plot.R
\name{Plot_Embed_Continous}
\alias{Plot_Embed_Continous}
\title{Title Plot 2D embedding with specific gene expression level as color scale}
\usage{
Plot_Embed_Continous(exprmatx, embd, genename, title = "")
}
\arguments{
\item{exprmatx}{Dataframe dataframe with each row represents a gene and each column represents a cell}

\item{embd}{datafram. By calling Embedding(exprmatx).}

\item{genename}{Gene Symbol, eg "Lars2"}

\item{title}{title of plot}
}
\value{
a ggplot object
}
\description{
Title Plot 2D embedding with specific gene expression level as color scale
}
\examples{
lars2 <- Plot_Embed_Continous(exprmatx,embd,"Lars2")
}
