% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Plot.R
\name{Plot_QC_Features}
\alias{Plot_QC_Features}
\title{Title Box and dot plot of QC features.}
\usage{
Plot_QC_Features(sum, group, features)
}
\arguments{
\item{sum}{Dataframe. Summary table. CellxFeature,c("RawReads","CleanReads"...)}

\item{group}{List. pattern of group identifier (which will be used in group)}

\item{features}{List. Pattern of selected feature, allow fuzzy search.}
}
\value{
ggplot2 object list. equal length as features list.
}
\description{
Title Box and dot plot of QC features.
}
\examples{
sum <- read.table("/Analysis/bioinfo/wenxingzhao/project/21_RAW_time_series/2019_06_Plate1_Blank_0h_24h/summary/QC_summary/Merge_Summary.txt",header = T, row.names = 1,stringsAsFactors = F) \\cr plot_list <- Plot_QC_Features(sum,group = c("blank","0h","24h"),features = c("Clean","Protein"))
}
