% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Plot.R
\name{Plot_Raw_Clean_Reads}
\alias{Plot_Raw_Clean_Reads}
\title{Title}
\usage{
Plot_Raw_Clean_Reads(sum, col = c(1, 2), srt = T)
}
\arguments{
\item{sum}{Dataframe. Summary table. CellxFeature,c("RawReads","CleanReads"...)}

\item{col}{list (default col=c(1,2)). Indicates which columns represents c("RawReads","CleanReads")}

\item{srt}{Bool (default srt=TRUE). Whether sort the bar by Clean Reads.}
}
\value{
a ggplot object
}
\description{
Title
}
\examples{
sum <- read.table("/Analysis/bioinfo/wenxingzhao/project/21_RAW_time_series/2019_06_Plate1_Blank_0h_24h/summary/QC_summary/Merge_Summary.txt",header = T, row.names = 1,stringsAsFactors = F)\\n p <- Plot_Raw_Clean_Reads(sum = sum,srt = F)
}
