% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Utils.R
\name{qqnorm}
\alias{qqnorm}
\title{Title Quantile normalization for single cell expression profile}
\usage{
qqnorm(x)
}
\arguments{
\item{x}{expression matrix row is gene col is cell}
}
\value{
Normalized expression profile
}
\description{
Title Quantile normalization for single cell expression profile
}
\examples{
qqnorm(TPM_matrix)
}
