% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Summary.R
\name{TopCells}
\alias{TopCells}
\title{Title Returen Cells expressed highest number of genes}
\usage{
TopCells(COUNT, top = 10)
}
\arguments{
\item{COUNT}{a dataframe with each row represents a gene and each column represents a cell}

\item{top}{int(default 10) number of top cells to inspect}
}
\value{
DataFrame with Cell names, expressed gene numbers, gene dropout rate
}
\description{
Title Returen Cells expressed highest number of genes
}
\examples{
tc <- TopCells(ExprMatx)
}
